Completely changed app structure in new branch
As of this commit, this does not work. Hence it is in a new branch.
