+++ lib/pages/apphelp_page.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:howth_golf_live/static/app_drawer.dart';
 import 'package:howth_golf_live/static/app_resources.dart';
 
 class AppHelpPage extends StatelessWidget with AppResources {
@@ -12,8 +13,14 @@ class AppHelpPage extends StatelessWidget with AppResources {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-        drawer: elementBuilder.buildDrawer(context),
-        appBar: elementBuilder.buildAppBar(context, 'App Help'),
+        drawer: AppDrawer(),
+        appBar: AppBar(
+          centerTitle: true,
+          title: Text(constants.appHelpText,
+              style: TextStyle(color: appTheme.primaryColorDark)),
+          backgroundColor: appTheme.primaryColor,
+          iconTheme: IconThemeData(color: appTheme.primaryColorDark),
+        ),
         body: Center(child: Text('App Help page')));
   }
 }
