+++ lib/pages/clublinks_page.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:howth_golf_live/static/app_drawer.dart';
 import 'package:howth_golf_live/static/app_resources.dart';
 
 class ClubLinksPage extends StatelessWidget with AppResources {
@@ -12,8 +13,14 @@ class ClubLinksPage extends StatelessWidget with AppResources {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-        drawer: elementBuilder.buildDrawer(context),
-        appBar: elementBuilder.buildAppBar(context, constants.clubLinksText),
+        drawer: AppDrawer(),
+        appBar: AppBar(
+          centerTitle: true,
+          title: Text(constants.clubLinksText,
+              style: TextStyle(color: appTheme.primaryColorDark)),
+          backgroundColor: appTheme.primaryColor,
+          iconTheme: IconThemeData(color: appTheme.primaryColorDark),
+        ),
         body: Center(child: Text('Club Links page')));
   }
 }
