+++ lib/pages/home_page.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:howth_golf_live/static/app_flashing_element.dart';
 import 'package:howth_golf_live/static/app_resources.dart';
 
 class HomePage extends StatelessWidget with AppResources {
@@ -20,7 +21,7 @@ class HomePage extends StatelessWidget with AppResources {
             Padding(padding: EdgeInsetsDirectional.only(top: 90)),
             new Image.asset('images/icon.png'),
             Padding(padding: EdgeInsetsDirectional.only(top: 400)),
-            elementBuilder.buildFlashingElement(Text(
+            FlashingElement(Text(
               'Tap anywhere to begin!',
               style: appTheme.textTheme.subhead,
             ))
