+++ lib/pages/results_page.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:howth_golf_live/static/app_bar_complex.dart';
 import 'package:howth_golf_live/static/app_resources.dart';
 
 class ResultsPage extends StatefulWidget with AppResources {
@@ -12,20 +13,50 @@ class ResultsPage extends StatefulWidget with AppResources {
 }
 
 class _ResultsPageState extends State<ResultsPage> with AppResources {
+  /// Use once the backend is complete to request a list of results
+/*     
+import 'package:WEB SCRAPING PACKAGE';
+
+void _getResults() async {
+      
+      BUILD SMART WEB SCRAPER HERE
+
+    } */
+  List _getResults() {
+    List tempList = ['results abby', 'your', 'resultssss'];
+    return tempList;
+  }
+
+  Widget _buildResultsList(
+      String _searchText, List filteredResults, List results) {
+    if (!(_searchText.isEmpty)) {
+      List tempList = new List();
+      for (int i = 0; i < filteredResults.length; i++) {
+        if (filteredResults[i]
+            .toLowerCase()
+            .contains(_searchText.toLowerCase())) {
+          tempList.add(filteredResults[i]);
+        }
+      }
+      filteredResults = tempList;
+    }
+    return ListView.builder(
+      itemCount: results == null ? 0 : filteredResults.length,
+      itemBuilder: (BuildContext context, int index) {
+        return new ListTile(
+          title: Text(filteredResults[index]),
+          onTap: () => print(filteredResults[index]),
+        );
+      },
+    );
+  }
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
         body: DefaultTabController(
             length: 3,
-            child: Scaffold(
-                drawer: elementBuilder.buildDrawer(context),
-                appBar: elementBuilder.buildTabAppBar(context, 'Results'),
-                body: TabBarView(
-                  children: <Widget>[
-                    Text('Recent results!'),
-                    Text('Archived results!'),
-                    Text('Your favourite results!')
-                  ],
-                ))));
+            child: ComplexAppBar(_getResults, _buildResultsList,
+                title: constants.competitionsText)));
   }
 }
