+++ lib/static/app_flashing_element.dart
